+++ src/components/EChartsEditor/EChartsEditor.tsx
@@ -1,7 +1,8 @@
 import React from 'react';
 import { StandardEditorProps } from '@grafana/data';
-import { CodeEditor, Monaco } from '@grafana/ui';
-import { CodeLanguage, Format } from '../../constants';
+import { getTemplateSrv } from '@grafana/runtime';
+import { CodeEditor, CodeEditorSuggestionItem, CodeEditorSuggestionItemKind } from '@grafana/ui';
+import { CodeEditorSuggestions, CodeLanguage, Format } from '../../constants';
 
 /**
  * Monaco
@@ -18,9 +19,14 @@ interface Props extends StandardEditorProps {}
  */
 export const EChartsEditor: React.FC<Props> = ({ value, onChange, context }) => {
   /**
+   * Template Service to get Variables
+   */
+  const templateSrv = getTemplateSrv();
+
+  /**
    * Format On Mount
    */
-  const onEditorMount = (editor: monacoType.editor.IStandaloneCodeEditor, monaco: Monaco) => {
+  const onEditorMount = (editor: monacoType.editor.IStandaloneCodeEditor) => {
     if (context.options.editor.format !== Format.AUTO) {
       return;
     }
@@ -28,6 +34,24 @@ export const EChartsEditor: React.FC<Props> = ({ value, onChange, context }) => 
     setTimeout(() => {
       editor.getAction('editor.action.formatDocument').run();
     }, 100);
+  };
+
+  /**
+   * Suggestions
+   */
+  const getSuggestions = (): CodeEditorSuggestionItem[] => {
+    /**
+     * Add Variables
+     */
+    const suggestions = templateSrv.getVariables().map((variable) => {
+      return {
+        label: `\$\{${variable.name}\}`,
+        kind: CodeEditorSuggestionItemKind.Property,
+        detail: variable.description ? variable.description : variable.label,
+      };
+    });
+
+    return [...CodeEditorSuggestions, ...suggestions];
   };
 
   /**
@@ -43,14 +67,18 @@ export const EChartsEditor: React.FC<Props> = ({ value, onChange, context }) => 
       <CodeEditor
         language={CodeLanguage.JAVASCRIPT}
         showLineNumbers={true}
-        showMiniMap={true}
+        showMiniMap={(value && value.length) > 100}
         value={value}
         height={`${context.options.editor.height}px`}
         onBlur={(code) => {
           onChange(code);
         }}
+        onSave={(code) => {
+          onChange(code);
+        }}
         monacoOptions={monacoOptions}
         onEditorDidMount={onEditorMount}
+        getSuggestions={getSuggestions}
       />
     </div>
   );