+++ src/constants/editor.ts
@@ -1,3 +1,5 @@
+import { CodeEditorSuggestionItem, CodeEditorSuggestionItemKind } from '@grafana/ui';
+
 /**
  * Supported Languages
  */
@@ -20,4 +22,45 @@ export enum Format {
 export const FormatOptions = [
   { value: Format.NONE, label: 'Disabled' },
   { value: Format.AUTO, label: 'Auto' },
+];
+
+/**
+ * Suggestions
+ */
+export const CodeEditorSuggestions: CodeEditorSuggestionItem[] = [
+  {
+    label: 'data',
+    kind: CodeEditorSuggestionItemKind.Property,
+    detail: 'Result set of panel queries.',
+  },
+  {
+    label: 'theme',
+    kind: CodeEditorSuggestionItemKind.Property,
+    detail: 'Theme object.',
+  },
+  {
+    label: 'echartsInstance',
+    kind: CodeEditorSuggestionItemKind.Property,
+    detail: 'Instance of the ECharts.',
+  },
+  {
+    label: 'echarts',
+    kind: CodeEditorSuggestionItemKind.Property,
+    detail: 'ECharts library.',
+  },
+  {
+    label: 'ecStat',
+    kind: CodeEditorSuggestionItemKind.Property,
+    detail: 'A statistical and data mining tool.',
+  },
+  {
+    label: 'locationService',
+    kind: CodeEditorSuggestionItemKind.Property,
+    detail: 'Browser location and history.',
+  },
+  {
+    label: 'replaceVariables',
+    kind: CodeEditorSuggestionItemKind.Method,
+    detail: 'Interpolate variables.',
+  },
 ];